Update to 0.1.2 TPI fix

+++ tests/testthat/test_supported_saga_versions.R
@@ -46,73 +46,7 @@ testthat::test_that("initiation of SAGA-GIS 2.3.1", {
     saga <- saga_gis(saga_bin)
     testthat::expect_false(is.null(saga))
     testthat::expect_gt(length(saga), 0)
-    testthat::expect_is(
-      saga$grid_calculus$random_terrain(iterations = 1, radius = 1),
-      "RasterLayer")
-    unlink(dirname(saga_bin), recursive = TRUE)
-  }
-})
-
-
-testthat::test_that("initiation of SAGA-GIS 3.0.0", {
-  testthat::skip_on_cran()
-  testthat::skip_if_not(
-    Sys.info()["sysname"] == "Windows" & 
-      Sys.info()["machine"] == "x86-64")
-
-  saga_url <-
-    "https://sourceforge.net/projects/saga-gis/files/SAGA%20-%203/SAGA%20-%203.0.0/saga_3.0.0_x64.zip"
-  saga_bin <- download_saga(saga_url)
-  
-  if (!is.null(saga_bin)) {
-    saga <- saga_gis(saga_bin)
-    testthat::expect_false(is.null(saga))
-    testthat::expect_gt(length(saga), 0)
     testthat::expect_s4_class(
-      saga$grid_calculus$random_terrain(iterations = 1, radius = 1),
-      "RasterLayer")
-    unlink(dirname(saga_bin), recursive = TRUE)
-  }
-})
-
-
-testthat::test_that("initiation of SAGA-GIS 4.0.0", {
-  testthat::skip_on_cran()
-  testthat::skip_if_not(
-    Sys.info()["sysname"] == "Windows" & 
-      Sys.info()["machine"] == "x86-64")
-    
-  saga_url <-
-    "https://sourceforge.net/projects/saga-gis/files/SAGA%20-%204/SAGA%20-%204.0.0/saga_4.0.0_x64.zip"
-  saga_bin <- download_saga(saga_url)
-  
-  if (!is.null(saga_bin)) {
-    saga <- saga_gis(saga_bin)
-    testthat::expect_false(is.null(saga))
-    testthat::expect_gt(length(saga), 0)
-    testthat::expect_is(
-      saga$grid_calculus$random_terrain(iterations = 1, radius = 1),
-      "RasterLayer")
-    unlink(dirname(saga_bin), recursive = TRUE)
-  }
-})
-
-
-testthat::test_that("initiation of SAGA-GIS 5.0.0", {
-  testthat::skip_on_cran()
-  testthat::skip_if_not(
-    Sys.info()["sysname"] == "Windows" & 
-      Sys.info()["machine"] == "x86-64")
-    
-  saga_url <-
-    "https://sourceforge.net/projects/saga-gis/files/SAGA%20-%205/SAGA%20-%205.0.0/saga-5.0.0_x64.zip"
-  saga_bin <- download_saga(saga_url)
-  
-  if (!is.null(saga_bin)) {
-    saga <- saga_gis(saga_bin)
-    testthat::expect_false(is.null(saga))
-    testthat::expect_gt(length(saga), 0)
-    testthat::expect_is(
       saga$grid_calculus$random_terrain(iterations = 1, radius = 1),
       "RasterLayer")
     unlink(dirname(saga_bin), recursive = TRUE)
@@ -134,7 +68,7 @@ testthat::test_that("initiation of SAGA-GIS 6.0.0", {
     saga <- saga_gis(saga_bin)
     testthat::expect_false(is.null(saga))
     testthat::expect_gt(length(saga), 0)
-    testthat::expect_is(
+    testthat::expect_s4_class(
       saga$grid_calculus$random_terrain(iterations = 1, radius = 1),
       "RasterLayer")
     unlink(dirname(saga_bin), recursive = TRUE)
@@ -156,7 +90,7 @@ testthat::test_that("initiation of SAGA-GIS 7.0.0", {
     saga <- saga_gis(saga_bin)
     testthat::expect_false(is.null(saga))
     testthat::expect_gt(length(saga), 0)
-    testthat::expect_is(
+    testthat::expect_s4_class(
       saga$grid_calculus$random_terrain(iterations = 1, radius = 1),
       "RasterLayer")
     unlink(dirname(saga_bin), recursive = TRUE)
@@ -178,7 +112,29 @@ testthat::test_that("initiation of SAGA-GIS 7.4.0", {
     saga <- saga_gis(saga_bin)
     testthat::expect_false(is.null(saga))
     testthat::expect_gt(length(saga), 0)
-    testthat::expect_is(
+    testthat::expect_s4_class(
+      saga$grid_calculus$random_terrain(iterations = 1, radius = 1),
+      "RasterLayer")
+    unlink(dirname(saga_bin), recursive = TRUE)
+  }
+})
+
+
+testthat::test_that("initiation of SAGA-GIS 8.0.1", {
+  testthat::skip_on_cran()
+  testthat::skip_if_not(
+    Sys.info()["sysname"] == "Windows" & 
+      Sys.info()["machine"] == "x86-64")
+  
+  saga_url <-
+    "https://sourceforge.net/projects/saga-gis/files/SAGA%20-%208/SAGA%20-%208.0.1/saga-8.0.1_x64.zip"
+  saga_bin <- download_saga(saga_url)
+  
+  if (!is.null(saga_bin)) {
+    saga <- saga_gis(saga_bin)
+    testthat::expect_false(is.null(saga))
+    testthat::expect_gt(length(saga), 0)
+    testthat::expect_s4_class(
       saga$grid_calculus$random_terrain(iterations = 1, radius = 1),
       "RasterLayer")
     unlink(dirname(saga_bin), recursive = TRUE)
